fix(user-service): change id from Int to Long

+++ user-service/src/main/kotlin/services/UserService.kt
@@ -6,8 +6,8 @@ import com.lorenzoog.gitkib.userservice.entities.User
 
 interface UserService {
   suspend fun findPaginated(page: Int, size: Int = 15): Page<User>
-  suspend fun findById(id: Int): User
-  suspend fun updateById(id: Int, updateDto: UserUpdateDto): User
+  suspend fun findById(id: Long): User
+  suspend fun updateById(id: Long, updateDto: UserUpdateDto): User
   suspend fun create(data: User.() -> Unit)
-  suspend fun deleteById(id: Int)
+  suspend fun deleteById(id: Long)
 }
